+++ examples/docker/kafka-oauth-strimzi/kafka/start.sh
@@ -20,10 +20,11 @@ else
   cp $SERVER_PROPERTIES_FILE /tmp/strimzi.properties
 fi
 
-if [[ "$1" == "--kraft" ]]; then
-  KAFKA_CLUSTER_ID="$(/opt/kafka/bin/kafka-storage.sh random-uuid)"
-  /opt/kafka/bin/kafka-storage.sh format -t $KAFKA_CLUSTER_ID -c /tmp/strimzi.properties
-fi
+KAFKA_CLUSTER_ID="$(/opt/kafka/bin/kafka-storage.sh random-uuid)"
+/opt/kafka/bin/kafka-storage.sh format -t $KAFKA_CLUSTER_ID -c /tmp/strimzi.properties \
+    --add-scram 'SCRAM-SHA-512=[name=admin,password=admin-secret]' \
+    --add-scram 'SCRAM-SHA-512=[name=alice,password=alice-secret]'
+echo "Initialised kafka storage for KRaft and added user secrets for SCRAM"
 
 # add Strimzi kafka-oauth-* jars and their dependencies to classpath
 export CLASSPATH="/opt/kafka/libs/strimzi/*:$CLASSPATH"