+++ testsuite/docker/kafka/scripts/start_no_wait.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 set -e
 
-[ "$KAFKA_ZOOKEEPER_CONNECT" == "" ] && KAFKA_ZOOKEEPER_CONNECT=localhost:2181
-[ "$KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS" == "" ] && KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS=6000
-
 ./simple_kafka_config.sh | tee /tmp/strimzi.properties
+echo "Config created"
+
+KAFKA_CLUSTER_ID="$(/opt/kafka/bin/kafka-storage.sh random-uuid)"
+/opt/kafka/bin/kafka-storage.sh format -t $KAFKA_CLUSTER_ID -c /tmp/strimzi.properties \
+    --add-scram 'SCRAM-SHA-512=[name=admin,password=admin-secret]' \
+    --add-scram 'SCRAM-SHA-512=[name=alice,password=alice-secret]'
+echo "Initialised kafka storage for KRaft"
 
 # set log dir to writable directory
 if [ "$LOG_DIR" == "" ]; then
@@ -25,10 +29,10 @@ echo "CLASSPATH=$CLASSPATH"
 if [ "$PROMETHEUS_AGENT_CONFIG" == "" ]; then
 
   if [ "$PROMETHEUS_AGENT_VERSION" == "" ]; then
-    PROMETHEUS_AGENT_VERSION=$(ls /opt/kafka/libs/jmx_prometheus* | sed -E -n 's/.*([0-9]+\.[0-9]+\.[0-9]+).*$/\1/p')
+    PROMETHEUS_AGENT_VERSION=$(ls /opt/kafka/libs/strimzi/jmx_prometheus* | sed -E -n 's/.*([0-9]+\.[0-9]+\.[0-9]+).*$/\1/p')
   fi
 
-  export PROMETHEUS_AGENT_CONFIG="-javaagent:/opt/kafka/libs/jmx_prometheus_javaagent-$PROMETHEUS_AGENT_VERSION.jar=9404:/opt/kafka/config/strimzi/metrics-config.yml"
+  export PROMETHEUS_AGENT_CONFIG="-javaagent:/opt/kafka/libs/strimzi/jmx_prometheus_javaagent-$PROMETHEUS_AGENT_VERSION.jar=9404:/opt/kafka/config/strimzi/metrics-config.yml"
 fi
 echo "PROMETHEUS_AGENT_CONFIG=$PROMETHEUS_AGENT_CONFIG"