+++ testsuite/mockoauth-jdk17-tests/src/test/java/io/strimzi/kafka/oauth/server/authorizer/TestAuthzUtil.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2017-2023, Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.kafka.oauth.server.authorizer;
+
+public class TestAuthzUtil {
+
+    public static void clearKeycloakAuthorizerService() {
+        KeycloakAuthorizerService.clearInstance();
+    }
+}