+++ RELEASE_NOTES.md
@@ -1,6 +1,24 @@
 Release Notes
 =============
 
+0.16.0
+------
+
+### Using Kafka 4.0.0
+
+Kafka 4.0.0 server-side libraries are built with Java 17 bytecode compatibility. The client libraries are still built with Java 11 bytecode compatibility.
+
+### Java 17 required for building the project
+
+Java 17 is now required for building the project libraries. The example clients and the testsuite can also run with Java 11.
+All the components are built with Java 11 bytecode compatibility except `kafka-oauth-keycloak-authorizer` which requires Java 17 due to the dependency on server-side Kafka 4.0.0 libraries.
+
+### Removed support for KeycloakAuthorizer ACL delegation in Zookeeper mode
+
+`KeycloakAuthorizer` can be configured to delegate authorization decision to standard ACL authorizer provided by Kafka.
+Since Zookeeper mode is no longer supported, the ACL authorizer delegation only works if the Kafka node runs in KRaft mode.
+If `KeycloakAuthorizer` is deployed to Kafka running in Zookeeper mode, and `strimzi.authorization.delegate.to.kafka.acl` is set to `true`, the broker will fail to start.
+
 0.15.0
 ------